aaf: Implement Capture mode support
This patch implements the capture mode support from the AAF plugin.
Simply put, this mode works as follows: AVTP packets are received from
the network, the PCM samples are retrieved and presented to the alsa-lib
layer at the presentation time. In summary, the capture mode implements
a typical AVTP Listener.

Once the AAF device is put in running state, packet reception is
started. Every time an AVTP packet is received the plugin checks if it
is valid (according to the stream configuration provided by the user)
and copies the PCM samples to the audio buffer. Note that at this
moment, the samples are not presented to the alsa-lib layer (i.e.
hw_ptr is not updated).

When the first AVTP packet is received, the media clock is set to start
at the presentation time from that packet. At every tick from the media
clock, PCM samples are presented to the alsa-lib layer.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>